refactor: clean hooks up properly
